Fix `runWithDindSwarmAndRegistry` extract
